requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pymysql==1.1.0
cryptography==42.0.2
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.13.1
email-validator==2.1.0

# AI/ML Dependencies
sentence-transformers>=2.2.0
scikit-learn>=1.3.0
numpy>=1.24.0
torch>=2.0.0
